added auto conversion of change files
