+++ package.json
@@ -54,6 +54,7 @@
       ],
       "*.{js,jsx,json,scss,html,md}": [
         "node ./scripts/lint-staged/prettier",
+        "node ./scripts/lint-staged/auto-convert-change-files",
         "git add"
       ]
     },
